+++ packages/core/package.json
@@ -30,12 +30,15 @@
       ],
       "types": [
         "dist/types/index.d.ts"
+      ],
+      "types/internal": [
+        "dist/types/internal/index.d.ts"
       ]
     }
   },
   "packageManager": "yarn@3.1.1",
   "scripts": {
-    "build": "tsup src/index.ts src/fixtures/index.ts src/types/index.ts --format esm,cjs --dts",
+    "build": "tsup src/index.ts src/fixtures/index.ts src/types/index.ts src/types/internal/index.ts --format esm,cjs --dts",
     "dev": "yarn build --watch"
   },
   "devDependencies": {
