+++ packages/core/src/types/internal/internal.ts
@@ -1,0 +1,23 @@
+/**
+ * The `BlueprintStore` modifies the some types as below:
+ */
+
+import { Conjunction, Criterion } from "../blueprint";
+
+export type InternalCriterion = Criterion & {
+  id: Criterion["condition_id"];
+  uid: number;
+};
+
+export type InternalCriterionWithPosition = InternalCriterion & {
+  position: number;
+};
+
+export type InternalConjunction = Conjunction & {
+  id: undefined;
+  uid: number;
+};
+
+export type InternalBlueprint = Array<InternalCriterion | InternalConjunction>;
+
+export type GroupedBlueprint = InternalCriterionWithPosition[][];
