+++ packages/refine-react/src/index.ts
@@ -1,1 +1,2 @@
+export * from "refine-core/types";
 export * from "./tailwind";
