+++ packages/refine-react/src/tailwind/components/criterion/use-criterion.ts
@@ -1,5 +1,6 @@
 import { createContext, useContext } from "react";
-import { Criterion, InternalCriterionWithPosition } from "refine-core/types";
+import type { Criterion } from "refine-core/types";
+import type { InternalCriterionWithPosition } from "refine-core/types/internal";
 
 export type CriterionContext = InternalCriterionWithPosition & {
   updateCondition: (conditionId: Criterion["condition_id"]) => void;
